Crate next_permutation

A library to enumerate permutations, shuffles, and partitions.

Functions

for_each_pairing

Calls f for each pairing (ascending $(2,2,\dots,2)$-shuffle) of a in lexicographic order.

for_each_partition

Calls f for each partition of n in lexicographic order.

for_each_partition_rev

Calls f for each partition of n in the reverse of the lexicographic order.

for_each_permutation

Calls f for each permutation of a in lexicographic order.

for_each_shuffle

Calls f for each $(K, N - K)$-shuffle of a in lexicographic order.

next_pairing

Returns the next pairing (ascending $(2,2,\dots,2)$-shuffle) of a in lexicographic order.

next_partition

Returns the next partition of a in lexicographic order.

next_permutation

Returns the next permutation of a in lexicographic order.

next_shuffle

Returns the next $(K, N - K)$-shuffle of a in lexicographic order.

pairings

Returns all pairings (ascending $(2,2,\dots,2)$-shuffles) of a in lexicographic order.

partitions

Returns all partitions of n in lexicographic order.

partitions_rev

Returns all partitions of n in the reverse of the lexicographic order.

permutations

Returns all permutations of a in lexicographic order.

prev_partition

Returns the previous partition of a in lexicographic order.

shuffles

Returns all $(K, N - K)$-shuffles of a in lexicographic order.